+++ src/Menu/Menu.php
@@ -31,9 +31,11 @@ class Menu extends AbstractStaticPluginMainMenuProvider {
 	public function getStaticTopItems(): array {
 		return [
 			self::dic()->globalScreen()->mainmenu()->topLinkItem(self::dic()->globalScreen()->identification()->plugin(self::plugin()
-				->getPluginObject(), $this)->identifier(ilMailLoggerPlugin::PLUGIN_ID))->withTitle(ilMailLoggerPlugin::PLUGIN_NAME)
-				->withAction(self::dic()->ctrl()->getLinkTargetByClass([ ilUIPluginRouterGUI::class, LogGUI::class ], LogGUI::CMD_LOG))
-				->withAvailableCallable(function (): bool {
+				->getPluginObject(), $this)->identifier(ilMailLoggerPlugin::PLUGIN_ID))->withTitle(self::plugin()
+				->translate("log", LogGUI::LANG_MODULE_LOG))->withAction(self::dic()->ctrl()->getLinkTargetByClass([
+					ilUIPluginRouterGUI::class,
+					LogGUI::class
+				], LogGUI::CMD_LOG))->withAvailableCallable(function (): bool {
 					return self::plugin()->getPluginObject()->isActive();
 				})->withVisibilityCallable(function (): bool {
 					return self::access()->hasLogAccess();