+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.1.2]
+- Using some new ILIAS 5.3 UI's
+- Fix reset offset
+- Translate menu title
+
 ## [1.1.1]
 - Fix latest ILIAS 5.4